+++ shop/base/models/__init__.py
@@ -1,0 +1,14 @@
+__all__ = [
+    "Scale",
+    "Manufacturer",
+    "ScaleModel",
+    "User",
+    "Buy",
+]
+
+from base.models.scale import Scale
+from base.models.manufacturer import Manufacturer
+from base.models.scalemodel import ScaleModel
+from base.models.user import User
+from base.models.buy import Buy
+
